requirements.txt
python-dotenv
requests
pywhatkit
schedule
